rebase 17/1/25

+++ app/javascript/src/utilities/UserTemplatesHelpers.js
@@ -1,13 +1,14 @@
 import ProfilesFetcher from 'src/fetchers/ProfilesFetcher';
 import UsersFetcher from 'src/fetchers/UsersFetcher';
+
 const key = 'ketcher-tmpls';
 
 const createAddAttachmentidToNewUserTemplate = async (newValue, newItem, deleteIdx) => {
   const res = await ProfilesFetcher.uploadUserTemplates({
     content: JSON.stringify(newItem),
-  }).catch(err => console.log("err in create"));
+  }).catch((err) => console.log('err in create'));
   const attachment_id = res?.template_details?.filename;
-  newItem['props']['path'] = attachment_id;
+  newItem.props.path = attachment_id;
   newValue[newValue.length - 1] = newItem;
   if (deleteIdx) newValue.splice(deleteIdx, 1);
   window.removeEventListener('storage', null);
@@ -31,7 +32,7 @@ const removeUserTemplate = async (listOfLocalid, oldValue) => {
 
 const updateUserTemplateDetails = async (oldValue, newValue) => {
   const listOfLocalNames = newValue.map(
-    (item) => JSON.parse(item.struct).header.moleculeName
+    (item) => JSON.parse(item.struct)?.header?.moleculeName
   );
   for (let i = 0; i < oldValue.length; i++) {
     const localItem = JSON.parse(oldValue[i].struct);
@@ -39,9 +40,7 @@ const updateUserTemplateDetails = async (oldValue, newValue) => {
     if (!exists) {
       await ProfilesFetcher.deleteUserTemplate({
         path: oldValue[i].props.path,
-      }).catch(() =>
-        console.log('ISSUE WITH DELETE', localItem?.props?.path)
-      );
+      }).catch(() => console.log('ISSUE WITH DELETE', localItem?.props?.path));
       createAddAttachmentidToNewUserTemplate(newValue, newValue[i], i);
       break;
     }
@@ -50,12 +49,12 @@ const updateUserTemplateDetails = async (oldValue, newValue) => {
 
 const onEventListen = async (event) => {
   let { newValue, oldValue } = event;
-  if (newValue.length && oldValue.length) {
+  if (newValue && oldValue && newValue.length && oldValue.length) {
     newValue = JSON.parse(newValue);
     oldValue = JSON.parse(oldValue);
     if (event.key === key) { // matching key && deleteAllowed
       if (newValue.length > oldValue.length) { // when a new template is added
-        let newItem = newValue[newValue.length - 1];
+        const newItem = newValue[newValue.length - 1];
         createAddAttachmentidToNewUserTemplate(newValue, newItem);
       } else if (newValue.length < oldValue.length) { // when a template is deleted
         const listOfLocalid = newValue.map((item) => item.props.path);
